Moves StateViewModel to ui-core.
