+++ financial-connections/src/main/java/com/stripe/android/financialconnections/features/consent/ConsentPreviewParameterProvider.kt
@@ -12,7 +12,7 @@ import com.stripe.android.financialconnections.model.Image
 import com.stripe.android.financialconnections.model.LegalDetailsBody
 import com.stripe.android.financialconnections.model.LegalDetailsNotice
 import com.stripe.android.financialconnections.model.ServerLink
-import com.stripe.android.financialconnections.presentation.Async.Success
+import com.stripe.android.uicore.presentation.Async.Success
 import com.stripe.android.financialconnections.ui.theme.Theme
 
 internal class ConsentPreviewParameterProvider : PreviewParameterProvider<ConsentPreviewState> {
