+++ financial-connections/src/main/java/com/stripe/android/financialconnections/features/manualentry/ManualEntryPreviewParameterProvider.kt
@@ -4,10 +4,10 @@ import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import com.stripe.android.core.exception.APIException
 import com.stripe.android.financialconnections.R
 import com.stripe.android.financialconnections.features.manualentry.ManualEntryPreviewParameterProvider.PreviewState
-import com.stripe.android.financialconnections.presentation.Async.Fail
-import com.stripe.android.financialconnections.presentation.Async.Loading
-import com.stripe.android.financialconnections.presentation.Async.Success
-import com.stripe.android.financialconnections.presentation.Async.Uninitialized
+import com.stripe.android.uicore.presentation.Async.Fail
+import com.stripe.android.uicore.presentation.Async.Loading
+import com.stripe.android.uicore.presentation.Async.Success
+import com.stripe.android.uicore.presentation.Async.Uninitialized
 
 internal class ManualEntryPreviewParameterProvider : PreviewParameterProvider<PreviewState> {
     override val values = sequenceOf(
